Reorganized files in bug-serve to be more manageable
